refactor: migrate OTA firmware updates to tasks
